fix(plugin-graphql): stop importing bare zudoku in build-path code
