+++ packages/plugin-graphql/src/resolveEndpointUrl.ts
@@ -1,0 +1,12 @@
+import { joinUrl } from "zudoku";
+import { isSchemaUrl } from "./interfaces.js";
+
+export const resolveEndpointUrl = (
+  endpoint: string | undefined,
+  baseUrl: string | undefined,
+): string | undefined => {
+  if (!endpoint) return undefined;
+  if (isSchemaUrl(endpoint)) return endpoint;
+
+  return baseUrl ? joinUrl(baseUrl, endpoint) : joinUrl(endpoint);
+};
